added the principal folders and created the Home and Favorites pages
