--- src/App.js
@@ -1,13 +1,0 @@
-import './App.css';
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <h1>Hacker news</h1>
-      </header>
-    </div>
-  );
-}
-
-export default App;
