+++ src/pages/Favorites/index.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function Home() {
+  return (
+    <>
+      <h1>My favorite Hacker news list</h1>
+    </>
+  );
+}
