+++ src/pages/Home/index.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function Home() {
+  return (
+    <>
+      <h1>Hacker news list</h1>
+    </>
+  );
+}
